+++ src/index.ts
@@ -54,4 +54,5 @@ export type {
   BaseProps,
   BaseDateProps,
   BaseTimeProps,
+  PickerColors,
 } from './types';